Optimize touch state

+++ VoodooInput/VoodooInputSimulator/VoodooInputSimulatorDevice.hpp
@@ -124,7 +124,7 @@ private:
     VoodooInput* engine;
     AbsoluteTime start_timestamp;
     OSData* new_get_report_buffer = NULL;
-    UInt64 touch_state[15];
+    UInt8 touch_state[15];
     int stylus_check = 0;
     IOWorkLoop* work_loop;
     IOCommandGate* command_gate;
